pyproject.toml
[project]
name = "discord-name-changer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "translate>=3.6.1",
]
